+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## HEAD
+## 4.6.0 (2018-09-29)
 
 * Ensure URLs are properly encoded in `NetHttp#download` and `#open` using Addressable (@linyaoli)